--- a/pkg/ingester/ingester_test.go
+++ b/pkg/ingester/ingester_test.go
@@ -303,23 +303,24 @@ func TestIngesterAppendOutOfOrderAndDuplicate(t *testing.T) {
 	m := labelPairs{
 		{Name: model.MetricNameLabel, Value: "testmetric"},
 	}
-	ctx := context.Background()
-	err := ing.append(ctx, userID, m, 1, 0, client.API)
+	state := ing.userStates.getOrCreate(userID)
+	require.NotNil(t, state)
+	err := ing.append(state, m, 1, 0, client.API)
 	require.NoError(t, err)
 
 	// Two times exactly the same sample (noop).
-	err = ing.append(ctx, userID, m, 1, 0, client.API)
+	err = ing.append(state, m, 1, 0, client.API)
 	require.NoError(t, err)
 
 	// Earlier sample than previous one.
-	err = ing.append(ctx, userID, m, 0, 0, client.API)
+	err = ing.append(state, m, 0, 0, client.API)
 	require.Contains(t, err.Error(), "sample timestamp out of order")
 	errResp, ok := httpgrpc.HTTPResponseFromError(err)
 	require.True(t, ok)
 	require.Equal(t, errResp.Code, int32(400))
 
 	// Same timestamp as previous sample, but different value.
-	err = ing.append(ctx, userID, m, 1, 1, client.API)
+	err = ing.append(state, m, 1, 1, client.API)
 	require.Contains(t, err.Error(), "sample with repeated timestamp but different value")
 	errResp, ok = httpgrpc.HTTPResponseFromError(err)
 	require.True(t, ok)
@@ -337,7 +338,9 @@ func TestIngesterAppendBlankLabel(t *testing.T) {
 		{Name: "bar", Value: ""},
 	}
 	ctx := user.InjectOrgID(context.Background(), userID)
-	err := ing.append(ctx, userID, lp, 1, 0, client.API)
+	state := ing.userStates.getOrCreate(userID)
+	require.NotNil(t, state)
+	err := ing.append(state, lp, 1, 0, client.API)
 	require.NoError(t, err)
 
 	res, _, err := runTestQuery(ctx, t, ing, labels.MatchEqual, labels.MetricName, "testmetric")