--- a/pkg/ingester/ingester_test.go
+++ b/pkg/ingester/ingester_test.go
@@ -310,21 +310,18 @@ func TestIngesterAppendOutOfOrderAndDuplicate(t *testing.T) {
 
 	// Two times exactly the same sample (noop).
 	err = ing.append(state, m, 1, 0, client.API)
-	require.NoError(t, err)
+	require.Error(t, err)
+	mse, ok := err.(*memorySeriesError)
+	require.True(t, ok)
+	require.True(t, mse.noReport)
 
 	// Earlier sample than previous one.
 	err = ing.append(state, m, 0, 0, client.API)
 	require.Contains(t, err.Error(), "sample timestamp out of order")
-	errResp, ok := httpgrpc.HTTPResponseFromError(err)
-	require.True(t, ok)
-	require.Equal(t, errResp.Code, int32(400))
 
 	// Same timestamp as previous sample, but different value.
 	err = ing.append(state, m, 1, 1, client.API)
 	require.Contains(t, err.Error(), "sample with repeated timestamp but different value")
-	errResp, ok = httpgrpc.HTTPResponseFromError(err)
-	require.True(t, ok)
-	require.Equal(t, errResp.Code, int32(400))
 }
 
 // Test that blank labels are removed by the ingester