Found some (small) possible optimization while debugging the WAL code. Reuse of some objects created in `ingester.append` by creating them in `ingester.Push` and passing them on has some improvements in push time (no changes in allocs or B/op). Along with that, updating metrics only once for `Push` call instead of in each `append` call.
```
benchmark                                        old ns/op     new ns/op     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     15628904      14997580      -4.04%
BenchmarkIngesterPush/encoding=Varbit-8          15608147      14949577      -4.22%
BenchmarkIngesterPush/encoding=Bigchunk-8        15625913      15132356      -3.16%
```
If the readability if getting bad with this change, we can only reuse the user state for now.