@pracucci 2019-11-15
I took a brief look at the change set and I'm wondering if the complexity in the changes is enough to justify a small performance improvement.