@codesome 2019-11-18
I guess not. But I could change the code to only re-use the `state`, which has the maximum contribution to the reduced latency.